feat: 책별 대표 위치 등록 API 구현

+++ server/src/main/java/com/codecozy/server/service/BookService.java
@@ -1,10 +1,7 @@
 package com.codecozy.server.service;
 
 import com.codecozy.server.context.StatusCode;
-import com.codecozy.server.dto.request.ReactionCommentRequest;
-import com.codecozy.server.dto.request.ReportCommentRequest;
-import com.codecozy.server.dto.request.ReadingBookCreateRequest;
-import com.codecozy.server.dto.request.ReviewCreateRequest;
+import com.codecozy.server.dto.request.*;
 import com.codecozy.server.dto.response.DefaultResponse;
 import com.codecozy.server.entity.*;
 import com.codecozy.server.repository.*;
@@ -40,7 +37,7 @@ public class BookService {
 
         // 자주 사용되는 객체 재사용 용도로 선언
         Book book = new Book();
-        LocationList locationList = new LocationList();
+        LocationList locationList = null;
 
         // isbn을 이용해 책 등록이 중복되었는지 검사
         if (bookRepository.findByIsbn(isbn) == null) {
@@ -49,28 +46,30 @@ public class BookService {
             bookRepository.save(book);
         }
 
-        long latitude = Long.parseLong(request.mainLocation().latitude());
-        long longitude = Long.parseLong(request.mainLocation().longitude());
+        if (request.mainLocation() != null) {
+            long latitude = Long.parseLong(request.mainLocation().latitude());
+            long longitude = Long.parseLong(request.mainLocation().longitude());
 
-        // latitude, longitude를 이용해 이미 등록한 위치인지 검사
-        if (locationRepository.findByLatitudeAndLongitude(latitude, longitude) == null) {
-            // 등록하지 않은 위치면 새로 등록
-            locationList = LocationList.create(request.mainLocation().placeName(), request.mainLocation().address(), latitude, longitude);
-            locationRepository.save(locationList);
-        }
-
-        locationList = locationRepository.findByLatitudeAndLongitude(latitude, longitude);
-        // 최근 위치 검색 기록에 없으면 추가
-        if (memberLocationRepository.findByMemberAndLocationList(member, locationList) == null) {
-            // 레코드가 5개 이상인지 확인, 5개 이상이면 날짜 순으로 정렬 후 가장 오래된 레코드 삭제
-            if (memberLocationRepository.count() >= 5) {
-                List<MemberLocation> memberLocationList = memberLocationRepository.findAllByOrderByDateAsc();
-                memberLocationRepository.delete(memberLocationList.get(0));
+            // latitude, longitude를 이용해 이미 등록한 위치인지 검사
+            if (locationRepository.findByLatitudeAndLongitude(latitude, longitude) == null) {
+                // 등록하지 않은 위치면 새로 등록
+                locationList = LocationList.create(request.mainLocation().placeName(), request.mainLocation().address(), latitude, longitude);
+                locationRepository.save(locationList);
             }
 
-            // 검색 기록에 추가
-            MemberLocation memberLocation = MemberLocation.create(member, locationList, LocalDate.now().toString());
-            memberLocationRepository.save(memberLocation);
+            locationList = locationRepository.findByLatitudeAndLongitude(latitude, longitude);
+            // 최근 위치 검색 기록에 없으면 추가
+            if (memberLocationRepository.findByMemberAndLocationList(member, locationList) == null) {
+                // 레코드가 5개 이상인지 확인, 5개 이상이면 날짜 순으로 정렬 후 가장 오래된 레코드 삭제
+                if (memberLocationRepository.count() >= 5) {
+                    List<MemberLocation> memberLocationList = memberLocationRepository.findAllByOrderByDateAsc();
+                    memberLocationRepository.delete(memberLocationList.get(0));
+                }
+
+                // 검색 기록에 추가
+                MemberLocation memberLocation = MemberLocation.create(member, locationList, LocalDate.now().toString());
+                memberLocationRepository.save(memberLocation);
+            }
         }
 
         // memberId와 isbn을 이용해 사용자별 리뷰 등록 책이 중복되었는지 검사
@@ -82,7 +81,6 @@ public class BookService {
 
         // isbn으로 책 찾고, latitude와 longitude로 위치 찾아서 bookRecord 생성에 사용
         book = bookRepository.findByIsbn(isbn);
-        locationList = locationRepository.findByLatitudeAndLongitude(latitude, longitude);
 
         // 등록하지 않은 책이면 등록
         BookRecord bookRecord = BookRecord.create(member, book, request.readingStatus(), request.bookType(), locationList, request.isMine(), request.isHidden(), request.startDate(), request.recentDate());
@@ -241,6 +239,49 @@ public class BookService {
             BookReview bookReview = BookReview.create(member, book, request.comment());
             bookReviewRepository.save(bookReview);
         }
+
+        return new ResponseEntity<>(
+                DefaultResponse.from(StatusCode.OK, "성공"),
+                HttpStatus.OK);
+    }
+
+    // 책별 대표 위치 등록 API (주소 테이블에 추가, 해당 책에 대표 위치 등록)
+    public ResponseEntity<DefaultResponse> addMainLocation(String token, String isbn, LocationCreateRequest request) {
+        // 사용자 받아오기
+        Long memberId = tokenProvider.getMemberIdFromToken(token);
+        Member member = memberRepository.findByMemberId(memberId);
+
+        // isbn으로 책 검색
+        Book book = bookRepository.findByIsbn(isbn);
+
+        // 자주 사용하는 위도, 경도 변수 따로 저장
+        long latitude = Long.parseLong(request.latitude());
+        long longitude = Long.parseLong(request.longitude());
+
+        // 위도, 경도로 해당 주소 찾기
+        LocationList locationList = locationRepository.findByLatitudeAndLongitude(latitude, longitude);
+
+        // 등록되지 않은 주소면 새로 등록
+        if(locationList == null) {
+            locationList = LocationList.create(request.placeName(), request.address(), latitude, longitude);
+            locationRepository.save(locationList);
+            locationList = locationRepository.findByLatitudeAndLongitude(latitude, longitude);
+        }
+
+        // 사용자 독서노트 검색
+        BookRecord bookRecord = bookRecordRepository.findByMemberAndBook(member, book);
+
+        // 대표 위치가 없으면 위치 등록
+        if (bookRecord.getLocationList() == null) {
+            bookRecord.setLocationList(locationList);
+        }
+
+        return new ResponseEntity<>(
+                DefaultResponse.from(StatusCode.OK, "성공"),
+                HttpStatus.OK);
+    }
+
+    public ResponseEntity<DefaultResponse> addpersonalDictionary(String token, String isbn, PersonalDictionaryRequest request) {
 
         return new ResponseEntity<>(
                 DefaultResponse.from(StatusCode.OK, "성공"),
